feat: 더미데이터 api 전송

+++ app/src/main/java/com/flint/presentation/collectioncreate/model/CollectionContentUiModel.kt
@@ -4,7 +4,7 @@ import kotlinx.collections.immutable.persistentListOf
 
 //TODO: UiModel 삭제 필요
 data class CollectionContentUiModel(
-    val contentId: Long,
+    val contentId: Int,
     val imageUrl: String,
     val title: String,
     val director: String,
@@ -14,14 +14,14 @@ data class CollectionContentUiModel(
         val dummyContentList =
             persistentListOf(
                 CollectionContentUiModel(
-                    contentId = 1L,
+                    contentId = 1,
                     imageUrl = "",
                     title = "인터스텔라",
                     director = "크리스토퍼 놀란",
                     createdYear = "2014",
                 ),
                 CollectionContentUiModel(
-                    contentId = 2L,
+                    contentId = 2,
                     imageUrl = "",
                     title = "해리포터 불의 잔",
                     director = "마이크 뉴웰",
